pyproject.toml
[tool.poetry]
name = "rag-fusion-mcp"
version = "0.1.0"
description = "Dual-index RAG fusion system with MCP server integration."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
  { include = "indexer" },
  { include = "server" }
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
numpy = "^1.26.4"
pandas = "^2.2.2"
PyMuPDF = "^1.24.10"
pdf2image = "^1.17.0"
pytesseract = "^0.3.10"
Pillow = "^10.4.0"
camelot-py = "^0.11.0"
ghostscript = "^0.8.1"
opencv-python = "^4.10.0.84"
scikit-learn = "^1.5.1"
sentence-transformers = "^2.7.0"
rank-bm25 = "^0.2.2"
faiss-cpu = "^1.8.0.post1"
typer = "^0.12.4"
rich = "^13.7.1"
tqdm = "^4.66.4"
fastapi = "^0.114.0"
uvicorn = { extras = ["standard"], version = "^0.30.1" }
pydantic = "^2.7.4"
cachetools = "^5.3.3"
slowapi = "^0.1.9"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
pytest-cov = "^5.0.0"
mypy = "^1.10.0"
ruff = "^0.5.5"
types-tqdm = "^4.66.0.20240417"
types-pytz = "^2024.1.0.20240417"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
